requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
pytest-xdist==3.5.0
pytest-rerunfailures==13.0
python-dotenv==1.0.0
allure-python-commons==2.13.2